bus/hp_dio/98643a: convert to anonymous namespace (nw)
