+++ src/devices/bus/hp_dio/hp98643.cpp
@@ -13,9 +13,80 @@
 //#define VERBOSE 1
 #include "logmacro.h"
 
-DEFINE_DEVICE_TYPE(HPDIO_98643, bus::hp_dio::dio16_98643_device, "dio98643", "HP98643A LANIC Ethernet card")
 
-namespace bus::hp_dio {
+namespace {
+
+static constexpr int REG_SWITCHES_REMOTE = 0x80;
+
+static constexpr int REG_SWITCHES_SELECT_CODE_MASK = 0x1f;
+static constexpr int REG_SWITCHES_SELECT_CODE_SHIFT = 0x00;
+
+static constexpr int REG_SWITCHES_INT_LEVEL_MASK = 0x03;
+static constexpr int REG_SWITCHES_INT_LEVEL_SHIFT = 0x05;
+
+static constexpr uint16_t REG_ID = 0x15;
+
+static constexpr uint16_t REG_STATUS_ACK = 0x04;
+
+static constexpr uint16_t REG_SC_REV = 0x01;
+static constexpr uint16_t REG_SC_LOCK = 0x08;
+static constexpr uint16_t REG_SC_IP = 0x40;
+static constexpr uint16_t REG_SC_IE = 0x80;
+
+class dio16_98643_device :
+		public device_t,
+		public bus::hp_dio::device_dio16_card_interface
+{
+public:
+	// construction/destruction
+	dio16_98643_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock);
+
+protected:
+	dio16_98643_device(const machine_config &mconfig, device_type type, const char *tag, device_t *owner, uint32_t clock);
+
+
+	// device-level overrides
+	virtual void device_start() override;
+	virtual void device_reset() override;
+
+	virtual ioport_constructor device_input_ports() const override;
+	virtual void device_add_mconfig(machine_config &config) override;
+
+private:
+	required_device<am7990_device> m_lance;
+	required_ioport m_switches;
+
+	void addrmap(address_map &map);
+	void update_int();
+	int get_irq_line();
+
+	uint16_t sc_r();
+	void sc_w(uint16_t data);
+	uint16_t id_r();
+	void id_w(uint16_t data);
+	uint16_t novram_r(offs_t offset);
+	void novram_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+
+	void lance_int_w(int state);
+	void lance_dma_out(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint16_t lance_dma_in(offs_t offset);
+
+	uint16_t m_novram[32] = {
+		0xfff0, 0xfff0, 0xfff0, 0xfff0,
+		0xfff0, 0xfff8, 0xfff0, 0xfff0,
+		0xfff0, 0xfff9, 0xfff0, 0xfff6,
+		0xfffa, 0xfff1, 0xffff, 0xfff1,
+		0xfff0, 0xfff0, 0xfff0, 0xfff0,
+		0xfff0, 0xfff0, 0xfff0, 0xfff0,
+		0xfff0, 0xfff0, 0xfff0, 0xfff0,
+		0xfff0, 0xfff0, 0xfffa, 0xfff9,
+	};
+
+	std::array<uint16_t, 8192> m_ram;
+
+	uint16_t m_sc;
+	bool m_installed_io;
+};
 
 void dio16_98643_device::device_add_mconfig(machine_config &config)
 {
@@ -212,4 +283,6 @@ void dio16_98643_device::addrmap(address_map &map)
 	map(0xc000, 0xffff).rw(FUNC(dio16_98643_device::novram_r), FUNC(dio16_98643_device::novram_w));
 }
 
-} // namespace bus::hp_dio
+} // anonymous namespace
+
+DEFINE_DEVICE_TYPE_PRIVATE(HPDIO_98643, bus::hp_dio::device_dio16_card_interface, dio16_98643_device, "dio98643", "HP98643A LANIC Ethernet card")
